fixed the bugs that marked in previous version

+++ probability_wrapper.cpp
@@ -28,17 +28,19 @@ void ProbabilityWrapper::set_banner_operator_num(unsigned int _banner_operator_n
 
 // calculate the initial threshold of getting a any star 6 operator
 // based on the range of uniform int distribution
-double ProbabilityWrapper::calc_init_star6_threshold(const unsigned int dist_left_border,
-                                                     const unsigned int dist_right_border) {
-  return (dist_right_border - dist_left_border + 1) * base_star6_rate;
+unsigned long long int ProbabilityWrapper::calc_init_star6_threshold(
+    const unsigned int dist_left_border, const unsigned int dist_right_border) {
+  return static_cast<unsigned long long int>(
+      (dist_right_border - dist_left_border + 1) * base_star6_rate);
 }
 
 // calculate the initial threshold of getting a specific star 6 operator
 // based on the range of uniform int distribution
-double ProbabilityWrapper::calc_init_target_star6_threshold(const unsigned int dist_left_border,
-                                                            const unsigned int dist_right_border) {
-  return (dist_right_border - dist_left_border + 1) * base_star6_rate *
-         on_banner_star6_conditional_rate / banner_operator_num;
+unsigned long long int ProbabilityWrapper::calc_init_target_star6_threshold(
+    const unsigned int dist_left_border, const unsigned int dist_right_border) {
+  return static_cast<unsigned long long int>(
+      (dist_right_border - dist_left_border + 1) * base_star6_rate *
+      on_banner_star6_conditional_rate / banner_operator_num);
 }
 
 // calculate the change step of the threshold of getting a any star 6 operator
